change package in all classes affected by the package rename
